Adm-897 [frontend]: hide retry button (#1429)
* ADM-879-hide-retry feat: hide retry button

* ADM-879-hide-retry feat: add image into trivyignore
