+++ .trivyignore
@@ -16,5 +16,6 @@ CVE-2024-22259
 CVE-2024-28085
 CVE-2024-22262
 CVE-2024-2961
+CVE-2024-33599
 CVE-2019-10744
 CVE-2024-33599
